+++ firmware/hdl/gt_mp7_core/gtl_fdl_wrapper/gtl/muon_comparators.vhd
@@ -3,6 +3,8 @@
 -- Comparators for transverse momentum, pseudorapidity, azimuth angle, quality and isolation of muon objects
 
 -- Version history:
+-- HB 2020-06-08: inserted comparators for "unconstraint pt" and "impact parameter" of new muon structure.
+-- HB 2019-06-14: updated for "five eta cuts". Used phi_windows_comp.
 -- HB 2019-06-14: updated for "five eta cuts". Used phi_windows_comp.
 -- HB 2019-05-06: renamed from muon_comparators_v2 to muon_comparators.
 -- HB 2015-09-24: renamed to "muon_comparators_v2" for removing "d_s_i" from generic. These constants used directly from gtl.pkg now.
@@ -38,7 +40,10 @@ entity muon_comparators is
         phi_w2_lower_limit : std_logic_vector;
         requested_charge: string(1 to 3);
         qual_lut : std_logic_vector;
-        iso_lut : std_logic_vector
+        iso_lut : std_logic_vector;
+        ptu_upper_limit : std_logic_vector;
+        ptu_lower_limit : std_logic_vector;
+        ip_lut : std_logic_vector        
     );
     port(
         data_i : in std_logic_vector;
@@ -56,6 +61,8 @@ architecture rtl of muon_comparators is
     signal qual : std_logic_vector(D_S_I_MUON.qual_high downto D_S_I_MUON.qual_low);
     signal iso : std_logic_vector(D_S_I_MUON.iso_high downto D_S_I_MUON.iso_low);
     signal charge : std_logic_vector(D_S_I_MUON.charge_high downto D_S_I_MUON.charge_low);
+    signal ptu : std_logic_vector(D_S_I_MUON.ptu_high downto D_S_I_MUON.ptu_low);
+    signal ip : std_logic_vector(D_S_I_MUON.ip_high downto D_S_I_MUON.ip_low);
 
     signal pt_comp : std_logic;
     signal eta_comp : std_logic;
@@ -63,6 +70,8 @@ architecture rtl of muon_comparators is
     signal qual_comp : std_logic;
     signal iso_comp : std_logic;
     signal charge_comp : std_logic;
+    signal ptu_comp : std_logic;
+    signal ip_comp : std_logic;
     
     signal no_muon : std_logic;
 
@@ -88,6 +97,11 @@ begin
 -- AND
 -- ISO LUT
 -- ************************************************
+-- HB 2020-06-08
+-- AND
+-- Pt unconstraint in range
+-- AND
+-- IP (impact parameter) LUT
 
     pt  <= data_i(D_S_I_MUON.pt_high downto D_S_I_MUON.pt_low);
     eta <= data_i(D_S_I_MUON.eta_high downto D_S_I_MUON.eta_low);
@@ -95,6 +109,8 @@ begin
     qual <= data_i(D_S_I_MUON.qual_high downto D_S_I_MUON.qual_low);
     iso <= data_i(D_S_I_MUON.iso_high downto D_S_I_MUON.iso_low);
     charge <= data_i(D_S_I_MUON.charge_high downto D_S_I_MUON.charge_low);
+    ptu <= data_i(D_S_I_MUON.ptu_high downto D_S_I_MUON.ptu_low);
+    ip <= data_i(D_S_I_MUON.ip_high downto D_S_I_MUON.ip_low);
     
 -- HB 2015-08-28: inserted check for "no muon" (all object parameters = 0)
     no_muon <= '1' when data_i = ZERO else '0';
@@ -156,7 +172,13 @@ begin
 -- Comparator for ISO bits with LUT
     iso_comp <= iso_lut(CONV_INTEGER(iso)); -- 4 bit LUT for isolation, because of 2 bits isolation
 
+-- Comparator for Pt unconstraint
+    ptu_comp <= '1' when (ptu >= ptu_lower_limit and ptu <= ptu_lower_limit) else '0';
+
+-- Comparator for IP bits with LUT
+    ip_comp <= ip_lut(CONV_INTEGER(ip)); -- 4 bit LUT for impact parameter, because of 2 bits impact parameter
+
 -- Comparators AND
-    comp_o <= pt_comp and eta_comp and phi_comp and qual_comp and iso_comp and charge_comp and not no_muon;
+    comp_o <= pt_comp and eta_comp and phi_comp and qual_comp and iso_comp and charge_comp and ptu_comp and ip_comp and not no_muon;
 
 end architecture rtl;