+++ firmware/hdl/gt_mp7_core/gtl_fdl_wrapper/gtl/muon_muon_correlation_condition.vhd
@@ -3,6 +3,7 @@
 -- Correlation Condition module for muon objects.
 
 -- Version history:
+-- HB 2020-06-09: implemented new muon structure with "unconstraint pt" and "impact parameter".
 -- HB 2019-06-17: updated for "five eta cuts".
 -- HB 2019-05-06: updated instances.
 -- HB 2019-05-06: renamed from muon_muon_correlation_condition_v4 to muon_muon_correlation_condition.
@@ -62,6 +63,9 @@ entity muon_muon_correlation_condition is
         requested_charge_muon1: string(1 to 3);
         qual_lut_muon1: std_logic_vector(2**(D_S_I_MUON_V2.qual_high-D_S_I_MUON_V2.qual_low+1)-1 downto 0);
         iso_lut_muon1: std_logic_vector(2**(D_S_I_MUON_V2.iso_high-D_S_I_MUON_V2.iso_low+1)-1 downto 0);
+        ptu_upper_limit_muon1: std_logic_vector(MAX_MUON_TEMPLATES_BITS-1 downto 0);
+        ptu_lower_limit_muon1: std_logic_vector(MAX_MUON_TEMPLATES_BITS-1 downto 0);
+        ip_lut_muon1: std_logic_vector(2**(D_S_I_MUON_V2.ip_high-D_S_I_MUON_V2.ip_low+1)-1 downto 0);
 
         muon2_object_low: natural;
         muon2_object_high: natural;
@@ -87,6 +91,9 @@ entity muon_muon_correlation_condition is
         requested_charge_muon2: string(1 to 3);
         qual_lut_muon2: std_logic_vector(2**(D_S_I_MUON_V2.qual_high-D_S_I_MUON_V2.qual_low+1)-1 downto 0);
         iso_lut_muon2: std_logic_vector(2**(D_S_I_MUON_V2.iso_high-D_S_I_MUON_V2.iso_low+1)-1 downto 0);
+        ptu_upper_limit_muon2: std_logic_vector(MAX_MUON_TEMPLATES_BITS-1 downto 0);
+        ptu_lower_limit_muon2: std_logic_vector(MAX_MUON_TEMPLATES_BITS-1 downto 0);
+        ip_lut_muon2: std_logic_vector(2**(D_S_I_MUON_V2.ip_high-D_S_I_MUON_V2.ip_low+1)-1 downto 0);
 
         requested_charge_correlation: string(1 to 2);
 
@@ -310,7 +317,10 @@ begin
                 phi_w2_lower_limit_muon1(D_S_I_MUON_V2.phi_high-D_S_I_MUON_V2.phi_low downto 0),
                 requested_charge_muon1,
                 qual_lut_muon1,
-                iso_lut_muon1
+                iso_lut_muon1,
+                ptu_upper_limit_muon1(D_S_I_MUON_V2.ptu_high-D_S_I_MUON_V2.ptu_low downto 0),
+                ptu_lower_limit_muon1(D_S_I_MUON_V2.ptu_high-D_S_I_MUON_V2.ptu_low downto 0),
+                ip_lut_muon1
             )
             port map(muon1_data_i(i), muon1_obj_vs_templ(i,1));
     end generate obj_templ1_l;
@@ -338,7 +348,10 @@ begin
                 phi_w2_lower_limit_muon2(D_S_I_MUON_V2.phi_high-D_S_I_MUON_V2.phi_low downto 0),
                 requested_charge_muon2,
                 qual_lut_muon2,
-                iso_lut_muon2
+                iso_lut_muon2,
+                ptu_upper_limit_muon2(D_S_I_MUON_V2.ptu_high-D_S_I_MUON_V2.ptu_low downto 0),
+                ptu_lower_limit_muon2(D_S_I_MUON_V2.ptu_high-D_S_I_MUON_V2.ptu_low downto 0),
+                ip_lut_muon2
             )
             port map(muon2_data_i(i), muon2_obj_vs_templ(i,1));
     end generate obj_templ2_l_l;