+++ CHANGELOG.md
@@ -4,6 +4,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
   
+## [v1.11.0] - 2020-06-09
+### Comment
+
+- mp7_ugt firmware release v1.11.0 (implemented new muon structure with "unconstraint pt" and "impact parameter").
+
+### Changed
+
+- source files:
+  - ../gtl/muon_comparators.vhd
+  - ../gtl/muon_obj_cuts.vhd
+  - ../gtl/muon_conditions.vhd
+  - ../gtl/muon_conditions_orm.vhd
+  - ../gtl/muon_muon_correlation_condition.vhd
+  - ../gtl/calo_muon_correlation_condition.vhd
+  - ../gtl/muon_esums_correlation_condition.vhd
+  - ../gtl/gtl_pkg.vhd
+  - ../gtl/gt_mp7_core_pkg.vhd
+
 ## [v1.10.7] - 2020-02-03
 ### Comment
 
@@ -14,7 +32,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - source files:
   - ../gtl/esums_comparators.vhd
   - ../gtl/min_bias_hf_conditions.vhd
-  
+
 ## [v1.10.6] - 2020-01-28
 ### Comment